DI components and modules was moved to packages.
